Implemented per region remember set
Added a global remember set in G1 for region remember sets to write
into. Updated G1 constrains to support write barriers. Added object
reference write barriers to G1Mutator. It updates the corresponding
remember set when when we have a reference write. After 256 writes
local remember set is saved to global and is flushed. The write
threshold should be an option and not hard wired. Added functionality
to check whether a region is eden, survivor or old. Added a counter
for survived GC of the region. Added some other helper functions to
Region and RegionList.